chore: refactor tests and reorganize processing mehtods into commit module
