+++ git-cliff-core/src/changelog.rs
@@ -2,11 +2,9 @@ use std::collections::HashMap;
 use std::io::{Read, Write};
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use log::{debug, trace};
-
-use crate::commit::Commit;
-use crate::config::{Config, GitConfig, ProcessingStep};
-use crate::error::{Error, Result};
+use crate::commit::process_commit_list;
+use crate::config::Config;
+use crate::error::Result;
 use crate::release::{Release, Releases};
 #[cfg(feature = "azure_devops")]
 use crate::remote::azure_devops::AzureDevOpsClient;
@@ -92,239 +90,6 @@ impl<'a> Changelog<'a> {
         Ok(())
     }
 
-    /// Checks the commits and returns an error if any unconventional commits
-    /// are found.
-    fn check_conventional_commits(commits: &Vec<Commit<'a>>) -> Result<()> {
-        log::debug!("Verifying that all commits are conventional");
-        let mut unconventional_count = 0;
-        commits.iter().for_each(|commit| {
-            if commit.conv.is_none() {
-                log::error!(
-                    "Commit {id} is not conventional:\n{message}",
-                    id = &commit.id[..7],
-                    message = commit
-                        .message
-                        .lines()
-                        .map(|line| { format!("    | {}", line.trim()) })
-                        .collect::<Vec<String>>()
-                        .join("\n")
-                );
-                unconventional_count += 1;
-            }
-        });
-
-        if unconventional_count > 0 {
-            return Err(Error::UnconventionalCommitsError(unconventional_count));
-        }
-
-        Ok(())
-    }
-
-    /// Splits the commits by their message lines.
-    /// Returns a new vector of commits with each line as a separate commit.
-    fn apply_split_commits(commits: &mut Vec<Commit<'a>>) -> Vec<Commit<'a>> {
-        let mut split_commits = Vec::new();
-        for commit in commits {
-            commit.message.lines().for_each(|line| {
-                let mut c = commit.clone();
-                c.message = line.to_string();
-                c.links.clear();
-                if !c.message.is_empty() {
-                    split_commits.push(c)
-                };
-            })
-        }
-        split_commits
-    }
-
-    /// Applies the commit parsers to the commits and returns the parsed
-    /// commits.
-    fn apply_commit_parsers(
-        commits: &mut Vec<Commit<'a>>,
-        git_config: &GitConfig,
-        summary: &mut Summary,
-    ) -> Vec<Commit<'a>> {
-        commits
-            .iter()
-            .filter_map(|commit| {
-                match commit.clone().parse(
-                    &git_config.commit_parsers,
-                    git_config.protect_breaking_commits,
-                    git_config.filter_commits,
-                ) {
-                    Ok(commit) => {
-                        summary.record_ok();
-                        Some(commit)
-                    }
-                    Err(e) => {
-                        summary.record_err(&e);
-                        Self::on_step_err(commit.clone(), e);
-                        None
-                    }
-                }
-            })
-            .collect::<Vec<_>>()
-    }
-
-    /// Applies the commit preprocessors to the commits and returns the
-    /// preprocessed commits.
-    fn apply_commit_preprocessors(
-        commits: &mut Vec<Commit<'a>>,
-        git_config: &GitConfig,
-        summary: &mut Summary,
-    ) -> Vec<Commit<'a>> {
-        commits
-            .iter()
-            .filter_map(|commit| {
-                // Apply commit parsers
-                match commit.clone().preprocess(&git_config.commit_preprocessors) {
-                    Ok(commit) => {
-                        summary.record_ok();
-                        Some(commit)
-                    }
-                    Err(e) => {
-                        summary.record_err(&e);
-                        Self::on_step_err(commit.clone(), e);
-                        None
-                    }
-                }
-            })
-            .collect::<Vec<_>>()
-    }
-
-    /// Converts the commits into conventional format if the configuration
-    /// requires it.
-    fn apply_into_conventional(
-        commits: &mut Vec<Commit<'a>>,
-        git_config: &GitConfig,
-        summary: &mut Summary,
-    ) -> Vec<Commit<'a>> {
-        commits
-            .iter()
-            .filter_map(|commit| {
-                let mut commit_into_conventional = Ok(commit.clone());
-                if git_config.conventional_commits {
-                    if !git_config.require_conventional &&
-                        git_config.filter_unconventional &&
-                        !git_config.split_commits
-                    {
-                        commit_into_conventional = commit.clone().into_conventional();
-                    } else if let Ok(conv_commit) = commit.clone().into_conventional() {
-                        commit_into_conventional = Ok(conv_commit);
-                    };
-                };
-                match commit_into_conventional {
-                    Ok(commit) => {
-                        summary.record_ok();
-                        Some(commit)
-                    }
-                    Err(e) => {
-                        summary.record_err(&e);
-                        Self::on_step_err(commit.clone(), e);
-                        None
-                    }
-                }
-            })
-            .collect::<Vec<_>>()
-    }
-
-    /// Applies the link parsers to the commits and returns the parsed commits.
-    fn apply_link_parsers(
-        commits: &mut Vec<Commit<'a>>,
-        git_config: &GitConfig,
-    ) -> Vec<Commit<'a>> {
-        commits
-            .iter()
-            .map(|commit| commit.clone().parse_links(&git_config.link_parsers))
-            .collect::<Vec<_>>()
-    }
-
-    /// Checks the commits and returns an error if any commits are not matched
-    /// by any commit parser.
-    fn check_unmatched_commits(commits: &Vec<Commit<'a>>) -> Result<()> {
-        log::debug!("Verifying that no commits are unmatched by commit parsers");
-        let mut unmatched_count = 0;
-        commits.iter().for_each(|commit| {
-            let is_unmatched = commit.group.is_none();
-            if is_unmatched {
-                log::error!(
-                    "Commit {id} was not matched by any commit parser:\n{message}",
-                    id = &commit.id[..7],
-                    message = commit
-                        .message
-                        .lines()
-                        .map(|line| { format!("    | {}", line.trim()) })
-                        .collect::<Vec<String>>()
-                        .join("\n")
-                );
-                unmatched_count += 1;
-            }
-        });
-
-        if unmatched_count > 0 {
-            return Err(Error::UnmatchedCommitsError(unmatched_count));
-        }
-
-        Ok(())
-    }
-
-    /// Processes the commit list based on the processing order defined in the
-    /// configuration.
-    fn process_commit_list(
-        commits: &mut Vec<Commit<'a>>,
-        git_config: &GitConfig,
-        summary: &mut Summary,
-    ) -> Result<()> {
-        for step in &git_config.processing_order.order {
-            match step {
-                ProcessingStep::CommitParsers => {
-                    debug!("Applying commit parsers...");
-                    *commits = Self::apply_commit_parsers(commits, git_config, summary);
-                }
-                ProcessingStep::CommitPreprocessors => {
-                    debug!("Applying commit preprocessors...");
-                    *commits = Self::apply_commit_preprocessors(commits, git_config, summary);
-                }
-                ProcessingStep::IntoConventional => {
-                    debug!("Converting commits to conventional format...");
-                    *commits = Self::apply_into_conventional(commits, git_config, summary);
-                }
-                ProcessingStep::LinkParsers => {
-                    debug!("Applying link parsers...");
-                    *commits = Self::apply_link_parsers(commits, git_config);
-                }
-                ProcessingStep::SplitCommits => {
-                    debug!("Splitting commits...");
-                    if git_config.split_commits {
-                        *commits = Self::apply_split_commits(commits);
-                    } else {
-                        debug!("Split commits is disabled, skipping...");
-                    }
-                }
-            }
-        }
-
-        if git_config.require_conventional {
-            Self::check_conventional_commits(commits)?;
-        }
-
-        if git_config.fail_on_unmatched_commit {
-            Self::check_unmatched_commits(commits)?;
-        }
-
-        Ok(())
-    }
-
-    /// Logs the error of a failed step from a single commit.
-    fn on_step_err(commit: Commit<'a>, error: Error) {
-        trace!(
-            "{} - {} ({})",
-            commit.id.chars().take(7).collect::<String>(),
-            error,
-            commit.message.lines().next().unwrap_or_default().trim()
-        );
-    }
-
     /// Processes the commits and omits the ones that doesn't match the
     /// criteria set by configuration file.
     fn process_commits(&mut self) -> Result<()> {
@@ -332,9 +97,9 @@ impl<'a> Changelog<'a> {
 
         let mut summary = Summary::default();
         for release in self.releases.iter_mut() {
-            Self::process_commit_list(&mut release.commits, &self.config.git, &mut summary)?;
+            process_commit_list(&mut release.commits, &self.config.git, &mut summary)?;
             for submodule_commits in release.submodule_commits.values_mut() {
-                Self::process_commit_list(submodule_commits, &self.config.git, &mut summary)?;
+                process_commit_list(submodule_commits, &self.config.git, &mut summary)?;
             }
         }
 
@@ -397,8 +162,8 @@ impl<'a> Changelog<'a> {
                 release
                     .previous
                     .as_ref()
-                    .and_then(|release| release.version.as_ref()) ==
-                    Some(skipped_tag)
+                    .and_then(|release| release.version.as_ref())
+                    == Some(skipped_tag)
             }) {
                 if let Some(previous_release) = self.releases.get_mut(release_index + 1) {
                     previous_release.previous = None;
@@ -427,10 +192,12 @@ impl<'a> Changelog<'a> {
     #[cfg(feature = "github")]
     fn get_github_metadata(&self, ref_name: Option<&str>) -> Result<crate::remote::RemoteMetadata> {
         use crate::remote::github;
-        if self.config.remote.github.is_custom ||
-            self.body_template
-                .contains_variable(github::TEMPLATE_VARIABLES) ||
-            self.footer_template
+        if self.config.remote.github.is_custom
+            || self
+                .body_template
+                .contains_variable(github::TEMPLATE_VARIABLES)
+            || self
+                .footer_template
                 .as_ref()
                 .is_some_and(|v| v.contains_variable(github::TEMPLATE_VARIABLES))
         {
@@ -477,10 +244,12 @@ impl<'a> Changelog<'a> {
     #[cfg(feature = "gitlab")]
     fn get_gitlab_metadata(&self, ref_name: Option<&str>) -> Result<crate::remote::RemoteMetadata> {
         use crate::remote::gitlab;
-        if self.config.remote.gitlab.is_custom ||
-            self.body_template
-                .contains_variable(gitlab::TEMPLATE_VARIABLES) ||
-            self.footer_template
+        if self.config.remote.gitlab.is_custom
+            || self
+                .body_template
+                .contains_variable(gitlab::TEMPLATE_VARIABLES)
+            || self
+                .footer_template
                 .as_ref()
                 .is_some_and(|v| v.contains_variable(gitlab::TEMPLATE_VARIABLES))
         {
@@ -534,10 +303,12 @@ impl<'a> Changelog<'a> {
     #[cfg(feature = "gitea")]
     fn get_gitea_metadata(&self, ref_name: Option<&str>) -> Result<crate::remote::RemoteMetadata> {
         use crate::remote::gitea;
-        if self.config.remote.gitea.is_custom ||
-            self.body_template
-                .contains_variable(gitea::TEMPLATE_VARIABLES) ||
-            self.footer_template
+        if self.config.remote.gitea.is_custom
+            || self
+                .body_template
+                .contains_variable(gitea::TEMPLATE_VARIABLES)
+            || self
+                .footer_template
                 .as_ref()
                 .is_some_and(|v| v.contains_variable(gitea::TEMPLATE_VARIABLES))
         {
@@ -587,10 +358,12 @@ impl<'a> Changelog<'a> {
         ref_name: Option<&str>,
     ) -> Result<crate::remote::RemoteMetadata> {
         use crate::remote::bitbucket;
-        if self.config.remote.bitbucket.is_custom ||
-            self.body_template
-                .contains_variable(bitbucket::TEMPLATE_VARIABLES) ||
-            self.footer_template
+        if self.config.remote.bitbucket.is_custom
+            || self
+                .body_template
+                .contains_variable(bitbucket::TEMPLATE_VARIABLES)
+            || self
+                .footer_template
                 .as_ref()
                 .is_some_and(|v| v.contains_variable(bitbucket::TEMPLATE_VARIABLES))
         {
@@ -638,10 +411,12 @@ impl<'a> Changelog<'a> {
         ref_name: Option<&str>,
     ) -> Result<crate::remote::RemoteMetadata> {
         use crate::remote::azure_devops;
-        if self.config.remote.azure_devops.is_custom ||
-            self.body_template
-                .contains_variable(azure_devops::TEMPLATE_VARIABLES) ||
-            self.footer_template
+        if self.config.remote.azure_devops.is_custom
+            || self
+                .body_template
+                .contains_variable(azure_devops::TEMPLATE_VARIABLES)
+            || self
+                .footer_template
                 .as_ref()
                 .is_some_and(|v| v.contains_variable(azure_devops::TEMPLATE_VARIABLES))
         {
@@ -885,9 +660,10 @@ mod test {
     use regex::Regex;
 
     use super::*;
-    use crate::commit::Signature;
+    use crate::commit::{Commit, Signature};
     use crate::config::{
-        Bump, ChangelogConfig, CommitParser, LinkParser, Remote, RemoteConfig, TextProcessor,
+        Bump, ChangelogConfig, CommitParser, GitConfig, LinkParser, ProcessingStep, Remote,
+        RemoteConfig, TextProcessor,
     };
 
     fn get_test_data() -> (Config, Vec<Release<'static>>) {
@@ -1289,24 +1065,29 @@ mod test {
             timestamp: Some(50_000_000),
             previous: None,
             repository: Some(String::from("/root/repo")),
-            submodule_commits: HashMap::from([(String::from("submodule_one"), vec![
-                Commit::new(
-                    String::from("sub0jkl12"),
-                    String::from("chore(app): submodule_one do nothing"),
-                ),
-                Commit::new(
-                    String::from("subqwerty"),
-                    String::from("chore: submodule_one <preprocess>"),
-                ),
-                Commit::new(
-                    String::from("subqwertz"),
-                    String::from("feat!: submodule_one support breaking commits"),
-                ),
-                Commit::new(
-                    String::from("subqwert0"),
-                    String::from("match(group): submodule_one support regex-replace for groups"),
-                ),
-            ])]),
+            submodule_commits: HashMap::from([(
+                String::from("submodule_one"),
+                vec![
+                    Commit::new(
+                        String::from("sub0jkl12"),
+                        String::from("chore(app): submodule_one do nothing"),
+                    ),
+                    Commit::new(
+                        String::from("subqwerty"),
+                        String::from("chore: submodule_one <preprocess>"),
+                    ),
+                    Commit::new(
+                        String::from("subqwertz"),
+                        String::from("feat!: submodule_one support breaking commits"),
+                    ),
+                    Commit::new(
+                        String::from("subqwert0"),
+                        String::from(
+                            "match(group): submodule_one support regex-replace for groups",
+                        ),
+                    ),
+                ],
+            )]),
             statistics: None,
             #[cfg(feature = "github")]
             github: crate::remote::RemoteReleaseMetadata {
@@ -1419,14 +1200,20 @@ mod test {
                 previous: Some(Box::new(test_release)),
                 repository: Some(String::from("/root/repo")),
                 submodule_commits: HashMap::from([
-                    (String::from("submodule_one"), vec![
-                        Commit::new(String::from("def349"), String::from("sub_one merge #4")),
-                        Commit::new(String::from("da8912"), String::from("sub_one merge #5")),
-                    ]),
-                    (String::from("submodule_two"), vec![Commit::new(
-                        String::from("ab76ef"),
-                        String::from("sub_two bump"),
-                    )]),
+                    (
+                        String::from("submodule_one"),
+                        vec![
+                            Commit::new(String::from("def349"), String::from("sub_one merge #4")),
+                            Commit::new(String::from("da8912"), String::from("sub_one merge #5")),
+                        ],
+                    ),
+                    (
+                        String::from("submodule_two"),
+                        vec![Commit::new(
+                            String::from("ab76ef"),
+                            String::from("sub_two bump"),
+                        )],
+                    ),
                 ]),
                 statistics: None,
                 #[cfg(feature = "github")]
